+++ VK_App/VK_App/Extension/PhotoCacheService+Extension.swift
@@ -1,0 +1,19 @@
+// PhotoCacheService+Extension.swift
+// Copyright © RoadMap. All rights reserved.
+
+import UIKit
+
+// MARK: - Расширение для PhotoCacheService
+
+extension PhotoCacheService {
+    class Table: DataReloadable {
+        let table: UITableView
+        init(table: UITableView) {
+            self.table = table
+        }
+
+        func reloadRow(atIndexpath indexPath: IndexPath) {
+            table.reloadRows(at: [indexPath], with: .none)
+        }
+    }
+}
